requirements.txt
aqualogic==3.4
paho-mqtt>=2.1
aiohttp>=3 # Required by aqualogic
websockets>=14 # Required by aqualogic